put the tests where they belong
